+++ tests/integration/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2013-2015 DataStax, Inc.
+# Copyright 2013-2016 DataStax, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -28,6 +28,7 @@ from cassandra.protocol import ConfigurationException
 
 try:
     from ccmlib.cluster import Cluster as CCMCluster
+    from ccmlib.dse_cluster import DseCluster
     from ccmlib.cluster_factory import ClusterFactory as CCMClusterFactory
     from ccmlib import common
 except ImportError as e:
@@ -82,38 +83,43 @@ USE_CASS_EXTERNAL = bool(os.getenv('USE_CASS_EXTERNAL', False))
 
 default_cassandra_version = '2.2.0'
 
-if USE_CASS_EXTERNAL:
-    if CCMClusterFactory:
-        # see if the external instance is running in ccm
-        path = common.get_default_path()
-        name = common.current_cluster_name(path)
-        CCM_CLUSTER = CCMClusterFactory.load(common.get_default_path(), name)
-        CCM_CLUSTER.start(wait_for_binary_proto=True, wait_other_notice=True)
 
-    # Not sure what's going on, but the server version query
-    # hangs in python3. This appears to be related to running inside of
-    # nosetests, and only for this query that would run while loading the
-    # module.
-    # This is a hack to make it run with default cassandra version for PY3.
-    # Not happy with it, but need to move on for now.
-    if not six.PY3:
-        cass_ver, _ = get_server_versions()
-        default_cassandra_version = '.'.join('%d' % i for i in cass_ver)
+def _get_cass_version_from_dse(dse_version):
+    if dse_version.startswith('4.6') or dse_version.startswith('4.5'):
+        cass_ver = "2.0"
+    elif dse_version.startswith('4.7') or dse_version.startswith('4.8'):
+        cass_ver = "2.1"
     else:
-        if not os.getenv('CASSANDRA_VERSION'):
-            log.warning("Using default C* version %s because external server cannot be queried" % default_cassandra_version)
+        log.error("Uknown dse version found {0}, defaulting to 2.1".format(dse_version))
+        cass_ver = "2.1"
 
+    return cass_ver
 
 CASSANDRA_DIR = os.getenv('CASSANDRA_DIR', None)
-CASSANDRA_VERSION = os.getenv('CASSANDRA_VERSION', default_cassandra_version)
+DSE_VERSION = os.getenv('DSE_VERSION', None)
+DSE_CRED = os.getenv('DSE_CREDS', None)
+if DSE_VERSION:
+    CASSANDRA_VERSION = _get_cass_version_from_dse(DSE_VERSION)
+else:
+    CASSANDRA_VERSION = os.getenv('CASSANDRA_VERSION', default_cassandra_version)
 
 CCM_KWARGS = {}
 if CASSANDRA_DIR:
     log.info("Using Cassandra dir: %s", CASSANDRA_DIR)
     CCM_KWARGS['install_dir'] = CASSANDRA_DIR
+
 else:
     log.info('Using Cassandra version: %s', CASSANDRA_VERSION)
     CCM_KWARGS['version'] = CASSANDRA_VERSION
+
+if DSE_VERSION:
+    log.info('Using DSE version: %s', DSE_VERSION)
+    if not CASSANDRA_DIR:
+        CCM_KWARGS['version'] = DSE_VERSION
+        if DSE_CRED:
+            log.info("Using DSE credentials file located at {0}".format(DSE_CRED))
+            CCM_KWARGS['dse_credentials_file'] = DSE_CRED
+
 
 if CASSANDRA_VERSION >= '2.2':
     default_protocol_version = 4
@@ -214,7 +220,11 @@ def use_cluster(cluster_name, nodes, ipformat=None, start=True):
             del tb
 
             log.debug("Creating new CCM cluster, {0}, with args {1}".format(cluster_name, CCM_KWARGS))
-            CCM_CLUSTER = CCMCluster(path, cluster_name, **CCM_KWARGS)
+            if DSE_VERSION:
+                log.error("creating dse cluster")
+                CCM_CLUSTER = DseCluster(path, cluster_name, **CCM_KWARGS)
+            else:
+                CCM_CLUSTER = CCMCluster(path, cluster_name, **CCM_KWARGS)
             CCM_CLUSTER.set_configuration_options({'start_native_transport': True})
             if CASSANDRA_VERSION >= '2.2':
                 CCM_CLUSTER.set_configuration_options({'enable_user_defined_functions': True})