+++ CHANGELOG.rst
@@ -1,3 +1,17 @@
+3.1.0
+=====
+
+Features
+--------
+* Pass name of server auth class to AuthProvider (PYTHON-454)
+* Surface schema agreed flag for DDL statements (PYTHON-458)
+* Automatically convert float and int to Decimal on serialization (PYTHON-468)
+
+Bug Fixes
+---------
+* Bus error (alignment issues) when running cython on some ARM platforms (PYTHON-450)
+* Overflow when decoding large collections (cython) (PYTHON-459)
+
 3.0.0
 =====
 November 24, 2015